Refactor NewGroup.vue
